chore(size-guard): move DoR regression test to refinement_dor_status_tests.rs
refinement_cap_tests.rs crossed the size-guard byte cap (54814 > 51200)
with the new regression test. Relocate the test to a sibling module
wired via the same #[path] pattern, sharing the existing fixture helpers
(now pub(super)); refinement_cap_tests.rs returns to its pre-PR size.
